refactor parsers
