+++ rllm/parser/__init__.py
@@ -1,6 +1,15 @@
-from rllm.parser.tool_parser.qwen_tool_parser import QwenToolParser
-from rllm.parser.tool_parser.r1_tool_parser import R1ToolParser
-from rllm.parser.tool_parser.tool_parser_base import ToolParser
+from rllm.parser.chat_template_parser import ChatTemplateParser, DeepseekQwenChatTemplateParser, LlamaChatTemplateParser, QwenChatTemplateParser
+from rllm.parser.tool_parser import QwenToolParser, R1ToolParser, ToolParser
+
+__all__ = [
+    "ChatTemplateParser",
+    "DeepseekQwenChatTemplateParser",
+    "QwenChatTemplateParser",
+    "LlamaChatTemplateParser",
+    "ToolParser",
+    "R1ToolParser",
+    "QwenToolParser",
+]
 
 PARSER_REGISTRY = {
     "r1": R1ToolParser,
@@ -11,12 +20,3 @@ PARSER_REGISTRY = {
 def get_tool_parser(parser_name: str) -> type[ToolParser]:
     assert parser_name in PARSER_REGISTRY, f"Tool parser {parser_name} not found in {PARSER_REGISTRY}"
     return PARSER_REGISTRY[parser_name]
-
-
-__all__ = [
-    "R1ToolParser",
-    "QwenToolParser",
-    "ToolParser",
-    "get_tool_parser",
-    "PARSER_REGISTRY",
-]
