--- rllm/parser/tool_parser/__init__.py
@@ -1,5 +1,0 @@
-from rllm.parser.tool_parser.qwen_tool_parser import QwenToolParser
-from rllm.parser.tool_parser.r1_tool_parser import R1ToolParser
-from rllm.parser.tool_parser.tool_parser_base import ToolParser
-
-__all__ = ["R1ToolParser", "QwenToolParser", "ToolParser"]
